Add new Role plugin implementation that sets role as null
